Updated page header
Header is now a fixed element and will stay at the top of the page for easy access to navigation.
Page styling affected by this change has been updated.
Breadcrumb navigation now working properly. New functions now split and join paths and labels properly so that the user is navigated to the right destination for each breadcrumb.